# services/platform-api/openapi.json
{
  "components": {
    "schemas": {
...
      "GeneratedArtifactList": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/GeneratedArtifactResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "nextCursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nextcursor"
          }
        },
        "required": [
          "items"
        ],
        "title": "GeneratedArtifactList",
        "type": "object"
      },
      "GeneratedArtifactPurpose": {
        "enum": [
          "reconstruction",
          "summary_video"
        ],
        "title": "GeneratedArtifactPurpose",
        "type": "string"
      },
      "GeneratedArtifactResponse": {
        "additionalProperties": false,
        "properties": {
          "artifactId": {
            "title": "Artifactid",
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "filename": {
            "title": "Filename",
            "type": "string"
          },
          "jobId": {
            "title": "Jobid",
            "type": "string"
          },
          "manifest": {
            "additionalProperties": true,
            "title": "Manifest",
            "type": "object"
          },
          "mediaType": {
            "title": "Mediatype",
            "type": "string"
          },
          "objectKey": {
            "title": "Objectkey",
            "type": "string"
          },
          "ownerId": {
            "title": "Ownerid",
            "type": "string"
          },
          "purpose": {
            "$ref": "#/components/schemas/GeneratedArtifactPurpose"
          },
          "retentionExpiresAt": {
            "format": "date-time",
            "title": "Retentionexpiresat",
            "type": "string"
          },
          "sha256": {
            "pattern": "^[a-f0-9]{64}$",
            "title": "Sha256",
            "type": "string"
          },
          "sizeBytes": {
            "title": "Sizebytes",
            "type": "integer"
          }
        },
        "required": [
          "artifactId",
          "ownerId",
          "jobId",
          "purpose",
          "filename",
          "mediaType",
          "sha256",
          "sizeBytes",
          "objectKey",
          "manifest",
          "createdAt",
          "retentionExpiresAt"
        ],
        "title": "GeneratedArtifactResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Stoma3D Platform API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/internal/v2/assets/generated": {
      "post": {
        "operationId": "upload_generated_artifact_internal_v2_assets_generated_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-Stoma3D-Service",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Stoma3D-Service"
            }
          },
          {
            "in": "header",
            "name": "X-Stoma3D-Timestamp",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Stoma3D-Timestamp"
            }
          },
          {
            "in": "header",
            "name": "X-Stoma3D-Nonce",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Stoma3D-Nonce"
            }
          },
          {
            "in": "header",
            "name": "X-Stoma3D-Content-SHA256",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Stoma3D-Content-Sha256"
            }
          },
          {
            "in": "header",
            "name": "X-Stoma3D-Signature",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Stoma3D-Signature"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "properties": {
                  "artifact": {
                    "format": "binary",
                    "type": "string"
                  },
                  "metadata": {
                    "contentMediaType": "application/json",
                    "type": "string"
                  }
                },
                "required": [
                  "metadata",
                  "artifact"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeneratedArtifactResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload Generated Artifact",
        "tags": [
          "internal generated assets"
        ]
      }
    },
    "/internal/v2/assets/{asset_id}/content": {
      "get": {
        "operationId": "get_internal_capture_asset_content_internal_v2_assets__asset_id__content_get",
        "parameters": [
          {
            "in": "path",
            "name": "asset_id",
            "required": true,
            "schema": {
              "title": "Asset Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Stoma3D-Service",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Stoma3D-Service"
            }
          },
          {
            "in": "header",
            "name": "X-Stoma3D-Timestamp",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Stoma3D-Timestamp"
            }
          },
          {
            "in": "header",
            "name": "X-Stoma3D-Nonce",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Stoma3D-Nonce"
            }
          },
          {
            "in": "header",
            "name": "X-Stoma3D-Content-SHA256",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Stoma3D-Content-Sha256"
            }
          },
          {
            "in": "header",
            "name": "X-Stoma3D-Signature",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Stoma3D-Signature"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Internal Capture Asset Content",
        "tags": [
          "internal generated assets"
        ]
      }
    },
...
    "/v2/generated-artifacts": {
      "get": {
        "operationId": "list_generated_artifacts_v2_generated_artifacts_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "before",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Before"
            }
          },
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeneratedArtifactList"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Generated Artifacts",
        "tags": [
          "internal generated assets"
        ]
      }
    },
    "/v2/generated-artifacts/{artifact_id}": {
      "get": {
        "operationId": "get_generated_artifact_v2_generated_artifacts__artifact_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "artifact_id",
            "required": true,
            "schema": {
              "title": "Artifact Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeneratedArtifactResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Generated Artifact",
        "tags": [
          "internal generated assets"
        ]
      }
    },
    "/v2/generated-artifacts/{artifact_id}/content": {
      "get": {
        "operationId": "get_generated_artifact_content_v2_generated_artifacts__artifact_id__content_get",
        "parameters": [
          {
            "in": "path",
            "name": "artifact_id",
            "required": true,
            "schema": {
              "title": "Artifact Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Generated Artifact Content",
        "tags": [
          "internal generated assets"
        ]
      }
    },
...
  }
}